Add empty ap and tp compile
