+++ pkg/sql/compile/compile_test.go
@@ -69,7 +69,7 @@ func TestCompile(t *testing.T) {
 		err := c.Compile(tc.pn, nil, testPrint)
 		require.NoError(t, err)
 		c.GetAffectedRows()
-		err = c.Run(0)
+		err = c.Run()
 		require.NoError(t, err)
 	}
 }
