+++ pkg/sql/compile/ddl.go
@@ -1,0 +1,157 @@
+// Copyright 2021 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package compile
+
+import (
+	"fmt"
+
+	"github.com/matrixorigin/matrixone/pkg/compress"
+	"github.com/matrixorigin/matrixone/pkg/container/types"
+	"github.com/matrixorigin/matrixone/pkg/errno"
+	"github.com/matrixorigin/matrixone/pkg/pb/plan"
+	"github.com/matrixorigin/matrixone/pkg/sql/errors"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine"
+)
+
+func (s *Scope) CreateDatabase(snapshot engine.Snapshot, c *Compile) error {
+	dbName := s.Plan.GetDdl().GetCreateDatabase().GetDatabase()
+	if _, err := c.e.Database(c.ctx, dbName, snapshot); err == nil {
+		if s.Plan.GetDdl().GetCreateDatabase().GetIfNotExists() {
+			return nil
+		}
+		return errors.New(errno.SyntaxErrororAccessRuleViolation, fmt.Sprintf("database %s already exists", dbName))
+	}
+	return c.e.Create(c.ctx, dbName, snapshot)
+}
+
+func (s *Scope) DropDatabase(snapshot engine.Snapshot, c *Compile) error {
+	dbName := s.Plan.GetDdl().GetDropDatabase().GetDatabase()
+	if _, err := c.e.Database(c.ctx, dbName, snapshot); err != nil {
+		if s.Plan.GetDdl().GetDropDatabase().GetIfExists() {
+			return nil
+		}
+		return err
+	}
+	return c.e.Delete(c.ctx, dbName, snapshot)
+}
+
+func (s *Scope) CreateTable(snapshot engine.Snapshot, c *Compile) error {
+	qry := s.Plan.GetDdl().GetCreateTable()
+	// convert the plan's cols to the execution's cols
+	planCols := qry.GetTableDef().GetCols()
+	exeCols := planColsToExeCols(planCols)
+
+	// convert the plan's defs to the execution's defs
+	planDefs := qry.GetTableDef().GetDefs()
+	exeDefs := planDefsToExeDefs(planDefs)
+
+	dbName := c.db
+	if qry.GetDatabase() != "" {
+		dbName = qry.GetDatabase()
+	}
+	dbSource, err := c.e.Database(c.ctx, dbName, snapshot)
+	if err != nil {
+		return err
+	}
+	tblName := qry.GetTableDef().GetName()
+	if _, err := dbSource.Relation(c.ctx, tblName); err == nil {
+		if qry.GetIfNotExists() {
+			return nil
+		}
+		return errors.New(errno.SyntaxErrororAccessRuleViolation, fmt.Sprintf("table '%s' already exists", tblName))
+	}
+	return dbSource.Create(c.ctx, tblName, append(exeCols, exeDefs...))
+}
+
+func (s *Scope) DropTable(snapshot engine.Snapshot, c *Compile) error {
+	qry := s.Plan.GetDdl().GetDropTable()
+
+	dbName := qry.GetDatabase()
+	dbSource, err := c.e.Database(c.ctx, dbName, snapshot)
+	if err != nil {
+		if qry.GetIfExists() {
+			return nil
+		}
+		return err
+	}
+	tblName := qry.GetTable()
+	if _, err := dbSource.Relation(c.ctx, tblName); err != nil {
+		if qry.GetIfExists() {
+			return nil
+		}
+		return err
+	}
+	return dbSource.Delete(c.ctx, tblName)
+}
+
+func planDefsToExeDefs(planDefs []*plan.TableDef_DefType) []engine.TableDef {
+	exeDefs := make([]engine.TableDef, len(planDefs))
+	for i, def := range planDefs {
+		switch defVal := def.GetDef().(type) {
+		case *plan.TableDef_DefType_Pk:
+			exeDefs[i] = &engine.PrimaryIndexDef{
+				Names: defVal.Pk.GetNames(),
+			}
+		case *plan.TableDef_DefType_Idx:
+			exeDefs[i] = &engine.IndexTableDef{
+				ColNames: defVal.Idx.GetColNames(),
+				Name:     defVal.Idx.GetName(),
+			}
+		case *plan.TableDef_DefType_Properties:
+			properties := make([]engine.Property, len(defVal.Properties.GetProperties()))
+			for i, p := range defVal.Properties.GetProperties() {
+				properties[i] = engine.Property{
+					Key:   p.GetKey(),
+					Value: p.GetValue(),
+				}
+			}
+			exeDefs[i] = &engine.PropertiesDef{
+				Properties: properties,
+			}
+		}
+	}
+	return exeDefs
+}
+
+func planColsToExeCols(planCols []*plan.ColDef) []engine.TableDef {
+	exeCols := make([]engine.TableDef, len(planCols))
+	for i, col := range planCols {
+		var alg compress.T
+		switch col.Alg {
+		case plan.CompressType_None:
+			alg = compress.None
+		case plan.CompressType_Lz4:
+			alg = compress.Lz4
+		}
+		colTyp := col.GetTyp()
+		exeCols[i] = &engine.AttributeDef{
+			Attr: engine.Attribute{
+				Name: col.Name,
+				Alg:  alg,
+				Type: types.Type{
+					Oid:       types.T(colTyp.GetId()),
+					Width:     colTyp.GetWidth(),
+					Precision: colTyp.GetPrecision(),
+					Scale:     colTyp.GetScale(),
+					Size:      colTyp.GetSize(),
+				},
+				Default: planCols[i].GetDefault(),
+				Primary: col.GetPrimary(),
+				Comment: col.GetComment(),
+			},
+		}
+	}
+	return exeCols
+}
