+++ pkg/sql/compile/scope.go
@@ -18,22 +18,9 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/matrixorigin/matrixone/pkg/container/nulls"
-	"github.com/matrixorigin/matrixone/pkg/container/vector"
-	"github.com/matrixorigin/matrixone/pkg/sql/colexec"
-	"github.com/matrixorigin/matrixone/pkg/sql/parsers/tree"
-
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec/connector"
-	"github.com/matrixorigin/matrixone/pkg/sql/colexec/deletion"
-	"github.com/matrixorigin/matrixone/pkg/sql/colexec/insert"
-	"github.com/matrixorigin/matrixone/pkg/sql/colexec/update"
 
-	"github.com/matrixorigin/matrixone/pkg/errno"
-
-	"github.com/matrixorigin/matrixone/pkg/compress"
 	"github.com/matrixorigin/matrixone/pkg/container/batch"
-	"github.com/matrixorigin/matrixone/pkg/container/types"
-	"github.com/matrixorigin/matrixone/pkg/pb/plan"
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec/group"
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec/limit"
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec/merge"
@@ -45,585 +32,11 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec/offset"
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec/order"
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec/top"
-	"github.com/matrixorigin/matrixone/pkg/sql/errors"
-	y "github.com/matrixorigin/matrixone/pkg/sql/plan"
 	"github.com/matrixorigin/matrixone/pkg/vm"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine"
 	"github.com/matrixorigin/matrixone/pkg/vm/pipeline"
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
 )
-
-func (s *Scope) CreateDatabase(ts uint64, snapshot engine.Snapshot, engine engine.Engine) error {
-	ctx := context.TODO()
-	dbName := s.Plan.GetDdl().GetCreateDatabase().GetDatabase()
-	if _, err := engine.Database(ctx, dbName, snapshot); err == nil {
-		if s.Plan.GetDdl().GetCreateDatabase().GetIfNotExists() {
-			return nil
-		}
-		return errors.New(errno.SyntaxErrororAccessRuleViolation, fmt.Sprintf("database %s already exists", dbName))
-	}
-	return engine.Create(ctx, dbName, snapshot)
-}
-
-func (s *Scope) DropDatabase(ts uint64, snapshot engine.Snapshot, engine engine.Engine) error {
-	ctx := context.TODO()
-	dbName := s.Plan.GetDdl().GetDropDatabase().GetDatabase()
-	if _, err := engine.Database(ctx, dbName, snapshot); err != nil {
-		if s.Plan.GetDdl().GetDropDatabase().GetIfExists() {
-			return nil
-		}
-		return err
-	}
-	return engine.Delete(ctx, dbName, snapshot)
-}
-
-func (s *Scope) CreateTable(ts uint64, snapshot engine.Snapshot, engine engine.Engine, dbName string) error {
-	ctx := context.TODO()
-	qry := s.Plan.GetDdl().GetCreateTable()
-	// convert the plan's cols to the execution's cols
-	planCols := qry.GetTableDef().GetCols()
-	exeCols := planColsToExeCols(planCols)
-
-	// convert the plan's defs to the execution's defs
-	planDefs := qry.GetTableDef().GetDefs()
-	exeDefs := planDefsToExeDefs(planDefs)
-
-	if qry.GetDatabase() != "" {
-		dbName = qry.GetDatabase()
-	}
-	dbSource, err := engine.Database(ctx, dbName, snapshot)
-	if err != nil {
-		return err
-	}
-	tblName := qry.GetTableDef().GetName()
-	if _, err := dbSource.Relation(ctx, tblName); err == nil {
-		if qry.GetIfNotExists() {
-			return nil
-		}
-		return errors.New(errno.SyntaxErrororAccessRuleViolation, fmt.Sprintf("table '%s' already exists", tblName))
-	}
-	return dbSource.Create(ctx, tblName, append(exeCols, exeDefs...))
-}
-
-func (s *Scope) DropTable(ts uint64, snapshot engine.Snapshot, engine engine.Engine) error {
-	ctx := context.TODO()
-	qry := s.Plan.GetDdl().GetDropTable()
-
-	dbName := qry.GetDatabase()
-	dbSource, err := engine.Database(ctx, dbName, snapshot)
-	if err != nil {
-		if qry.GetIfExists() {
-			return nil
-		}
-		return err
-	}
-	tblName := qry.GetTable()
-	if _, err := dbSource.Relation(ctx, tblName); err != nil {
-		if qry.GetIfExists() {
-			return nil
-		}
-		return err
-	}
-	return dbSource.Delete(ctx, tblName)
-}
-
-func (s *Scope) CreateIndex(ts uint64, snapshot engine.Snapshot, engine engine.Engine) error {
-	return nil
-}
-
-func (s *Scope) DropIndex(ts uint64, snapshot engine.Snapshot, engine engine.Engine) error {
-	return nil
-}
-
-func (s *Scope) Delete(ts uint64, snapshot engine.Snapshot, c *Compile) (uint64, error) {
-	s.Magic = Merge
-	arg := s.Instructions[len(s.Instructions)-1].Arg.(*deletion.Argument)
-	arg.Ts = ts
-
-	ctx := context.TODO()
-	if arg.DeleteCtxs[0].CanTruncate {
-		return arg.DeleteCtxs[0].TableSource.Truncate(ctx)
-	}
-
-	if err := s.MergeRun(c); err != nil {
-		return 0, err
-	}
-	return arg.AffectedRows, nil
-}
-
-func (s *Scope) Insert(ts uint64, snapshot engine.Snapshot, c *Compile) (uint64, error) {
-	s.Magic = Merge
-	arg := s.Instructions[len(s.Instructions)-1].Arg.(*insert.Argument)
-	arg.Ts = ts
-	if err := s.MergeRun(c); err != nil {
-		return 0, err
-	}
-	return arg.Affected, nil
-}
-
-func (s *Scope) Update(ts uint64, snapshot engine.Snapshot, c *Compile) (uint64, error) {
-	s.Magic = Merge
-	arg := s.Instructions[len(s.Instructions)-1].Arg.(*update.Argument)
-	arg.Ts = ts
-	if err := s.MergeRun(c); err != nil {
-		return 0, err
-	}
-	return arg.AffectedRows, nil
-}
-
-func (s *Scope) InsertValues(ts uint64, snapshot engine.Snapshot, engine engine.Engine, proc *process.Process, stmt *tree.Insert) (uint64, error) {
-	p := s.Plan.GetIns()
-
-	ctx := context.TODO()
-
-	dbSource, err := engine.Database(ctx, p.DbName, snapshot)
-	if err != nil {
-		return 0, err
-	}
-	relation, err := dbSource.Relation(ctx, p.TblName)
-	if err != nil {
-		return 0, err
-	}
-
-	bat := makeInsertBatch(p)
-
-	if p.OtherCols != nil {
-		p.ExplicitCols = append(p.ExplicitCols, p.OtherCols...)
-	}
-
-	if err := fillBatch(bat, p, stmt.Rows.Select.(*tree.ValuesClause).Rows, proc); err != nil {
-		return 0, err
-	}
-	batch.Reorder(bat, p.OrderAttrs)
-	if err := relation.Write(ctx, bat); err != nil {
-		return 0, err
-	}
-
-	return uint64(len(p.Columns[0].Column)), nil
-}
-
-func fillBatch(bat *batch.Batch, p *plan.InsertValues, rows []tree.Exprs, proc *process.Process) error {
-	rowCount := len(p.Columns[0].Column)
-
-	tmpBat := batch.NewWithSize(0)
-	tmpBat.Zs = []int64{1}
-
-	for i, v := range bat.Vecs {
-		switch v.Typ.Oid {
-		case types.T_json:
-			vs := make([][]byte, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.(*types.Bytes).Data
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_bool:
-			vs := make([]bool, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]bool)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_int8:
-			vs := make([]int8, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]int8)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_int16:
-			vs := make([]int16, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]int16)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_int32:
-			vs := make([]int32, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]int32)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_int64:
-			vs := make([]int64, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]int64)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_uint8:
-			vs := make([]uint8, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]uint8)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_uint16:
-			vs := make([]uint16, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]uint16)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_uint32:
-			vs := make([]uint32, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]uint32)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_uint64:
-			vs := make([]uint64, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]uint64)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_float32:
-			vs := make([]float32, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]float32)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_float64:
-			vs := make([]float64, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]float64)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_char, types.T_varchar, types.T_blob:
-			vs := make([][]byte, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.(*types.Bytes).Data
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_date:
-			vs := make([]types.Date, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]types.Date)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_datetime:
-			vs := make([]types.Datetime, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]types.Datetime)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_timestamp:
-			vs := make([]types.Timestamp, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]types.Timestamp)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_decimal64:
-			vs := make([]types.Decimal64, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]types.Decimal64)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		case types.T_decimal128:
-			vs := make([]types.Decimal128, rowCount)
-			{
-				for j, expr := range p.Columns[i].Column {
-					vec, err := colexec.EvalExpr(tmpBat, proc, expr)
-					if err != nil {
-						return y.MakeInsertError(v.Typ.Oid, p.ExplicitCols[i], rows, i, j)
-					}
-					if nulls.Any(vec.Nsp) {
-						nulls.Add(v.Nsp, uint64(j))
-					} else {
-						vs[j] = vec.Col.([]types.Decimal128)[0]
-					}
-				}
-			}
-			if err := vector.Append(v, vs); err != nil {
-				return err
-			}
-		default:
-			return fmt.Errorf("data truncation: type of '%v' doesn't implement", v.Typ)
-		}
-	}
-	return nil
-}
-
-func makeInsertBatch(p *plan.InsertValues) *batch.Batch {
-	attrs := make([]string, 0, len(p.OrderAttrs))
-
-	for _, col := range p.ExplicitCols {
-		attrs = append(attrs, col.Name)
-	}
-	for _, col := range p.OtherCols {
-		attrs = append(attrs, col.Name)
-	}
-
-	bat := batch.New(true, attrs)
-	idx := 0
-	for _, col := range p.ExplicitCols {
-		bat.Vecs[idx] = vector.New(types.Type{Oid: types.T(col.Typ.GetId())})
-		idx++
-	}
-	for _, col := range p.OtherCols {
-		bat.Vecs[idx] = vector.New(types.Type{Oid: types.T(col.Typ.GetId())})
-		idx++
-	}
-
-	return bat
-}
-
-func planDefsToExeDefs(planDefs []*plan.TableDef_DefType) []engine.TableDef {
-	exeDefs := make([]engine.TableDef, len(planDefs))
-	for i, def := range planDefs {
-		switch defVal := def.GetDef().(type) {
-		case *plan.TableDef_DefType_Pk:
-			exeDefs[i] = &engine.PrimaryIndexDef{
-				Names: defVal.Pk.GetNames(),
-			}
-		case *plan.TableDef_DefType_Idx:
-			exeDefs[i] = &engine.IndexTableDef{
-				ColNames: defVal.Idx.GetColNames(),
-				Name:     defVal.Idx.GetName(),
-			}
-		case *plan.TableDef_DefType_Properties:
-			properties := make([]engine.Property, len(defVal.Properties.GetProperties()))
-			for i, p := range defVal.Properties.GetProperties() {
-				properties[i] = engine.Property{
-					Key:   p.GetKey(),
-					Value: p.GetValue(),
-				}
-			}
-			exeDefs[i] = &engine.PropertiesDef{
-				Properties: properties,
-			}
-		}
-	}
-	return exeDefs
-}
-
-func planColsToExeCols(planCols []*plan.ColDef) []engine.TableDef {
-	exeCols := make([]engine.TableDef, len(planCols))
-	for i, col := range planCols {
-		var alg compress.T
-		switch col.Alg {
-		case plan.CompressType_None:
-			alg = compress.None
-		case plan.CompressType_Lz4:
-			alg = compress.Lz4
-		}
-		colTyp := col.GetTyp()
-		exeCols[i] = &engine.AttributeDef{
-			Attr: engine.Attribute{
-				Name: col.Name,
-				Alg:  alg,
-				Type: types.Type{
-					Oid:       types.T(colTyp.GetId()),
-					Width:     colTyp.GetWidth(),
-					Precision: colTyp.GetPrecision(),
-					Scale:     colTyp.GetScale(),
-					Size:      colTyp.GetSize(),
-				},
-				Default: planCols[i].GetDefault(),
-				Primary: col.GetPrimary(),
-				Comment: col.GetComment(),
-			},
-		}
-	}
-	return exeCols
-}
 
 // PrintScope Print is to format scope list
 func PrintScope(prefix []byte, ss []*Scope) {
